chore: Fix import paths, etc, after rename and refactor of hook
