+++ src/navigators/LoggedInStack.tsx
@@ -16,14 +16,14 @@ import { usePendingInvite } from '../hooks/usePendingInvite';
 import { useSetUserProfileEffect } from '../hooks/useSetUserProfileEffect';
 import { useDeveloperConfig } from '../hooks/useDeveloperConfig';
 import { useJoinCircles } from '../hooks/Circles/useJoinCircles';
-import { useMessagingProfiles } from '../hooks/useMessagingProfiles';
+import { useProfilesForAllTiles } from '../hooks/useMessagingProfiles';
 
 export function LoggedInStack() {
   const Stack = createNativeStackNavigator<LoggedInRootParamList>();
   const { inviteParams } = usePendingInvite();
   useSetUserProfileEffect();
   // Fetch profiles early but don't wait for them
-  useMessagingProfiles();
+  useProfilesForAllTiles();
   const {
     account,
     isLoading: isLoadingAccount,
